Changed dim to a long long int because the warnings were annoying

+++ DensityMap/densityMap.cpp
@@ -1,6 +1,6 @@
 #include "densityMap.h"
 
-DensityMap::DensityMap(int dim) {
+DensityMap::DensityMap(long long int dim) {
 	this->dim = dim;
 	threshold = 0;
